refactoring day01: ultrasonic sensor
added modules
refactor names
[] benchmarks
