+++ Projects/01_AdventCode2021/advent_of_code/src/d01_sonar_sweep/input/01_sensor_measurements.txt
