--- Projects/01_AdventCode2021/advent_of_code/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod d01_sonar_sweep;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        let result = 2 + 2;
-        assert_eq!(result, 4);
-    }
-}
