+++ Projects/01_AdventCode2021/advent_of_code/src/main.rs
@@ -1,42 +1,22 @@
-use advent_of_code::d01_sonar_sweep;
+use d01_sonar_sweep::ultrasonic_sensor;
 use std::time::{ Instant};
+
+pub mod d01_sonar_sweep{
+    pub mod ultrasonic_sensor;
+}
+
 fn main() {
   day_01();
 }
 
 pub fn day_01() {
-    let mut start = Instant::now();
-    println!(
-        "status sonar: {:?}",
-        d01_sonar_sweep::sonar_sweep::depth_increase()
-    );
-    let mut duration = start.elapsed();
+    let start = Instant::now();
+    match  ultrasonic_sensor::depth_descent_slope() {
+        Ok(value) =>  println!("status sonar: {:?}", value),
+        Err(e) => println!("not okay: {:?}", e)
+    }
+    let duration = start.elapsed();
     println!("Time elapsed in expensive_function() is: {:?}", duration);
 
 
-    start = Instant::now();
-    println!(
-        "status sonar functional: {:?}",
-        d01_sonar_sweep::sonar_sweep::depth_increase_functional()
-    );
-    duration = start.elapsed();
-    println!("Time elapsed in expensive_function() is: {:?}", duration);
-
-
-    start = Instant::now();
-    println!(
-        "status sonar functional2: {:?}",
-        d01_sonar_sweep::sonar_sweep::depth_increase_functional2()
-    );
-    duration = start.elapsed();
-    println!("Time elapsed in expensive_function() is: {:?}", duration);
-
-
-    start = Instant::now();
-    println!(
-        "status sonar functional2 slide window: {:?}",
-        d01_sonar_sweep::sonar_sweep::dept_increase_slide_filter()
-    );
-    duration = start.elapsed();
-    println!("Time elapsed in expensive_function() is: {:?}", duration);
 }
